pr fixes

+++ data/src/main/java/com/microsoft/azure/kusto/data/Utils.java
@@ -24,11 +24,11 @@ import java.util.HashMap;
 
 class Utils {
 
-    static Results post(String url, String aadAccessToken, String payload, Long timeoutMs) throws DataServiceException, DataClientException {
+    static Results post(String url, String aadAccessToken, String payload, Integer timeoutMs) throws DataServiceException, DataClientException {
 
         HttpClient httpClient;
         if (timeoutMs != null) {
-            RequestConfig requestConfig = RequestConfig.custom().setSocketTimeout(timeoutMs.intValue()).build();
+            RequestConfig requestConfig = RequestConfig.custom().setSocketTimeout(timeoutMs).build();
             httpClient = HttpClientBuilder.create().useSystemProperties().setDefaultRequestConfig(requestConfig).build();
         } else {
             httpClient =  HttpClients.createSystem();
